requirements.txt
# Core dependencies
requests>=2.32.0
pyyaml>=6.0.0
python-dateutil>=2.8.0

# Academic sources
feedparser>=6.0.0
biopython>=1.83.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Export formats
openpyxl>=3.1.0
jinja2>=3.1.0

# CLI enhancements
rich>=13.0.0
click>=8.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0